backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.1
python-dotenv==1.1.0
langchain==0.3.24
langchain-community==0.3.23
langchain-chroma==0.2.2
langchain-huggingface==0.1.2
langchain-groq==0.3.2
chromadb==0.6.3
sentence-transformers==3.4.1
PyMuPDF==1.25.5
pydantic==2.11.3
tiktoken==0.9.0
python-multipart==0.0.20
